requirements.txt
aioredis @ file:///Users/tedchoi/Library/Caches/pypoetry/artifacts/57/b1/ed/7ef24be02f4faed2e69ac3c9230da6819de61a06b0a6ecc2d007c3e19e/aioredis-1.3.1-py3-none-any.whl
asgiref @ file:///Users/tedchoi/Library/Caches/pypoetry/artifacts/c3/92/1c/3d42ce7ad1e9e2aa1621c3fa8ebc5e21bd880a768bd691fb4f0356baa8/asgiref-3.4.1-py3-none-any.whl
async-timeout @ file:///Users/tedchoi/Library/Caches/pypoetry/artifacts/e1/e7/dc/347dacf16e20e4b15a992103281583f3a03db422172dec1c5f08d68c07/async_timeout-3.0.1-py3-none-any.whl
attrs @ file:///Users/tedchoi/Library/Caches/pypoetry/artifacts/6f/a9/ee/569c37f69a8c365ee41d2340aeac0214ee8c0086b8d8db43a21545204b/attrs-21.2.0-py2.py3-none-any.whl
autobahn @ file:///Users/tedchoi/Library/Caches/pypoetry/artifacts/70/0a/4b/6ea27fca7918aaa8316eca5d83f3a5b38f2c353b06183bdd57435790d9/autobahn-21.3.1-py2.py3-none-any.whl
Automat @ file:///Users/tedchoi/Library/Caches/pypoetry/artifacts/b8/e5/0e/e86ac85851dbf0701b64dc7a0c07c823d0194d7e31a24f1ba60bed622d/Automat-20.2.0-py2.py3-none-any.whl
black @ file:///Users/tedchoi/Library/Caches/pypoetry/artifacts/03/0e/f9/f10727f8b0ba7665a77feb04560102e8367927f9148882a69f195f2c3f/black-21.8b0-py3-none-any.whl
cffi @ file:///Users/tedchoi/Library/Caches/pypoetry/artifacts/9c/1e/cd/ab56fd529c22877a2cc0db66966967c9e0b5e7df0b484c80ed72bbe119/cffi-1.14.6.tar.gz
channels @ file:///Users/tedchoi/Library/Caches/pypoetry/artifacts/12/9c/e7/f4252a50e125ab7ee58d50982122a1590d34c915b752ba13e03eedadd1/channels-3.0.4-py3-none-any.whl
channels-redis @ file:///Users/tedchoi/Library/Caches/pypoetry/artifacts/53/a7/a4/53ee82d053ed37afd0b867db272b656713767836b8f9c11aace0eb9768/channels_redis-3.3.0-py3-none-any.whl
click @ file:///Users/tedchoi/Library/Caches/pypoetry/artifacts/ae/32/83/e159324c1bd58177322f4e45f598d500fe22544bff20f53f55cf749da8/click-8.0.1-py3-none-any.whl
constantly @ file:///Users/tedchoi/Library/Caches/pypoetry/artifacts/b1/88/48/9251c310b06614794944e121657b60bba5de1686d8c2dfeb79f6f80ad0/constantly-15.1.0-py2.py3-none-any.whl
cryptography @ file:///Users/tedchoi/Library/Caches/pypoetry/artifacts/2f/3c/1a/f460722bc7217d7a9aaf69053eefb5011b821639a3c945cf50597937ae/cryptography-3.4.8-cp36-abi3-macosx_11_0_arm64.whl
daphne @ file:///Users/tedchoi/Library/Caches/pypoetry/artifacts/31/da/f5/87c148f67573404dad7fcd2c73264a32e1ef99902c3750d8f9473eb69a/daphne-3.0.2-py3-none-any.whl
Django @ file:///Users/tedchoi/Library/Caches/pypoetry/artifacts/c3/8c/59/50a3c14099a0ad0818e40ea40af0446cd732882bbed86d63e115ffcf0d/Django-3.2.7-py3-none-any.whl
django-phonenumber-field @ file:///Users/tedchoi/Library/Caches/pypoetry/artifacts/02/57/1f/7ec62f9d307b43867834e395875d9e95e449bfe25a7301721d86308022/django_phonenumber_field-5.2.0-py3-none-any.whl
django-rest-knox @ file:///Users/tedchoi/Library/Caches/pypoetry/artifacts/43/0d/50/a1429a322091055bb7484506f71e0a6c108b1ec2e7df445657f1dabc4e/django_rest_knox-4.1.0-py3-none-any.whl
django-seed @ file:///Users/tedchoi/Library/Caches/pypoetry/artifacts/8d/b2/65/0f227f31b058bb52e07752ac0fc91cc80813fa8fd8c8161a9358c08917/django-seed-0.2.2.tar.gz
djangorestframework @ file:///Users/tedchoi/Library/Caches/pypoetry/artifacts/e1/57/21/bfde361af294b9c9cd77dcab0182e58c2938fcee29a4e5724ef4fed0e0/djangorestframework-3.12.4-py3-none-any.whl
Faker @ file:///Users/tedchoi/Library/Caches/pypoetry/artifacts/3b/c3/92/dcd4497e10f69cb9c38ad17c0f083934035acc24e6fa37867d2517773b/Faker-8.12.1-py3-none-any.whl
hiredis @ file:///Users/tedchoi/Library/Caches/pypoetry/artifacts/ed/9d/12/bbd2a02276663ba3736fd859fdf6b0dddb41f37565a83457a376f27ac1/hiredis-2.0.0.tar.gz
hyperlink @ file:///Users/tedchoi/Library/Caches/pypoetry/artifacts/ac/c9/85/b172f5d08e937f31bc0feffe4e2ff08bef15ff1e4d382616b9f7b174b0/hyperlink-21.0.0-py2.py3-none-any.whl
idna @ file:///Users/tedchoi/Library/Caches/pypoetry/artifacts/ff/0d/80/b45ae1ddba13a20e406682fd42ce6c201d9bedc5f592bee00cf799a1fb/idna-3.2-py3-none-any.whl
incremental @ file:///Users/tedchoi/Library/Caches/pypoetry/artifacts/03/0f/d8/1f9c21de086b17b28e0ed5914f384759067c88a524742b94e72eb07020/incremental-21.3.0-py2.py3-none-any.whl
msgpack @ file:///Users/tedchoi/Library/Caches/pypoetry/artifacts/9b/f3/03/88bb91d836c2da43b3b8c6b015b3f71e456c660175224e448b585c342b/msgpack-1.0.2.tar.gz
mypy-extensions @ file:///Users/tedchoi/Library/Caches/pypoetry/artifacts/92/45/bf/1807ce854ff668d92602207a37bfa9316def2a3f257bd03c4c5be4bc9b/mypy_extensions-0.4.3-py2.py3-none-any.whl
pathspec @ file:///Users/tedchoi/Library/Caches/pypoetry/artifacts/0a/67/21/620031f7f2938bad77b057a68a4c2a334542a7402a3214ed48690caf8c/pathspec-0.9.0-py2.py3-none-any.whl
Pillow @ file:///Users/tedchoi/Library/Caches/pypoetry/artifacts/1a/d8/7b/6d7ad3ec1941fc2990944350d0b6c1e1559e24d86f1433148429c89fbf/Pillow-8.3.2-cp39-cp39-macosx_11_0_arm64.whl
platformdirs @ file:///Users/tedchoi/Library/Caches/pypoetry/artifacts/b5/1c/75/af1b48d59847e9602d4314e30b5eda42b3adb4d3efd767a4eefa54076d/platformdirs-2.3.0-py3-none-any.whl
pyasn1 @ file:///Users/tedchoi/Library/Caches/pypoetry/artifacts/7b/3a/54/42ce43b579bda01b9d79022fb733811594441e7a32e9f9a5a98f672bdc/pyasn1-0.4.8-py2.py3-none-any.whl
pyasn1-modules @ file:///Users/tedchoi/Library/Caches/pypoetry/artifacts/dd/b8/4f/b56433e0354274a31074995e01b8671751e9f0ed0001f5254e5b03a54f/pyasn1_modules-0.2.8-py2.py3-none-any.whl
pycparser @ file:///Users/tedchoi/Library/Caches/pypoetry/artifacts/37/8e/5a/0ea4f84bc7f11e0e3468110efa2c7783241ea7eaa63a92a751de06f78f/pycparser-2.20-py2.py3-none-any.whl
PyJWT @ file:///Users/tedchoi/Library/Caches/pypoetry/artifacts/33/2d/3b/c07cd30707fdf78d1c8279fdbc13b6aca325b3091040eff97623d32f78/PyJWT-2.1.0-py3-none-any.whl
pyOpenSSL @ file:///Users/tedchoi/Library/Caches/pypoetry/artifacts/27/2a/6f/de8129dc1f8ba5d04f604e929a7f055c4ca7c08ceec8bf801182df8774/pyOpenSSL-20.0.1-py2.py3-none-any.whl
python-dateutil @ file:///Users/tedchoi/Library/Caches/pypoetry/artifacts/53/f8/2a/7d63ce15df7386e9536e83413453f8aa845b47fb425f05c4ca2fb231c3/python_dateutil-2.8.2-py2.py3-none-any.whl
pytz @ file:///Users/tedchoi/Library/Caches/pypoetry/artifacts/b0/a7/8d/54de3ab4d1ff29abbbca1e9ccbaefdc2a1b290138311b84f73bee16de1/pytz-2021.1-py2.py3-none-any.whl
regex @ file:///Users/tedchoi/Library/Caches/pypoetry/artifacts/b8/e0/b7/469d9ff9b0800ee491815ef1501c4adafbfca680d8f2b06948fd5101bd/regex-2021.8.28.tar.gz
service-identity @ file:///Users/tedchoi/Library/Caches/pypoetry/artifacts/b1/52/73/b4dbc1db0eb8a4a2e9faf4231a0f8e54b2478b020bfe5c18fdcca0cbd7/service_identity-21.1.0-py2.py3-none-any.whl
six @ file:///Users/tedchoi/Library/Caches/pypoetry/artifacts/08/9f/47/c16ae03035fc69eaf100ea39657a49baaeef714e25a52575710c34cd48/six-1.16.0-py2.py3-none-any.whl
sqlparse @ file:///Users/tedchoi/Library/Caches/pypoetry/artifacts/bc/af/02/8aa71b17fd39b26b858293918c5c6412f83768a0bfd0a4626f660aa5f8/sqlparse-0.4.1-py3-none-any.whl
text-unidecode @ file:///Users/tedchoi/Library/Caches/pypoetry/artifacts/34/f9/c2/484c44b08bab89d472229bbd257fcc1d1c6273ee027f01cb08c4e3c309/text_unidecode-1.3-py2.py3-none-any.whl
tomli @ file:///Users/tedchoi/Library/Caches/pypoetry/artifacts/7b/a4/22/559325a4271b73ba93b936f2b2953235fba10bd955a50a9c0323354809/tomli-1.2.1-py3-none-any.whl
Twisted @ file:///Users/tedchoi/Library/Caches/pypoetry/artifacts/df/c4/16/fdec6b8954e38f9cb4aa67824db626cb67a81e106f0e3643333b19cc96/Twisted-21.7.0-py3-none-any.whl
txaio @ file:///Users/tedchoi/Library/Caches/pypoetry/artifacts/fa/08/32/c860d7ebf60bd51dfd8454092f6b968fd7ad9145276e7c651b6c7e088b/txaio-21.2.1-py2.py3-none-any.whl
typing-extensions @ file:///Users/tedchoi/Library/Caches/pypoetry/artifacts/4e/df/c9/f8b546ab8ea4b39e1c4f91a8360f6091cc3497ef3f9eb8fa65c0375b03/typing_extensions-3.10.0.2-py3-none-any.whl
zope.interface @ file:///Users/tedchoi/Library/Caches/pypoetry/artifacts/4d/44/d2/fed94f65efb5032ce06d3f420443cbf0862b2ab8f039fa0a3c76231aad/zope.interface-5.4.0.tar.gz
